Created SQL Queries to retrieve all tables from given database(uri), columns, columns types from each table, and primary key/foreign keys from each table
