--- newServer/controllers/gqlController.js
@@ -1,9 +1,0 @@
-gqlController = {}
-
-gqlController.makeSchema = function() {
-
-}
-
-gqlController.makeResolver = function() {
-
-}
