+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "NODE_ENV=production nodemon ./server/server.js",
     "build": "NODE_ENV=production webpack",
-    "dev": "NODE_ENV=development nodemon ./server/server.js & NODE_ENV=development webpack serve --hot --open"
+    "dev": "NODE_ENV=development node ./server/server.js & NODE_ENV=development webpack serve"
   },
   "nodemonConfig": {
     "ignore": [
