+++ server/sqlQuery.sql
@@ -69,3 +69,49 @@
 -- -- 
 -- -- 
 -- -- 
+
+--foreign keys on table
+--kcu: key column usage 
+--rel_tco: table constraints
+--rco: referential_constraint
+select kcu.table_name as foreign_table,
+          '>-' as rel,
+          rel_tco.table_name as primary_table,
+          string_agg(kcu.column_name, ', ') as fk_columns,
+          kcu.constraint_name
+   from information_schema.table_constraints tco
+   join information_schema.key_column_usage kcu
+             on tco.constraint_schema = kcu.constraint_schema
+             and tco.constraint_name = kcu.constraint_name
+   join information_schema.referential_constraints rco
+             on tco.constraint_schema = rco.constraint_schema
+             and tco.constraint_name = rco.constraint_name
+   join information_schema.table_constraints rel_tco
+             on rco.unique_constraint_schema = rel_tco.constraint_schema
+             and rco.unique_constraint_name = rel_tco.constraint_name
+   where tco.constraint_type = 'FOREIGN KEY'
+   group by kcu.table_schema,
+            kcu.table_name,
+            rel_tco.table_name,
+            rel_tco.table_schema,
+            kcu.constraint_name
+   order by kcu.table_schema,
+            kcu.table_name
+
+
+--Primary key query
+select kcu.table_schema,
+       kcu.table_name,
+       tco.constraint_name,
+       string_agg(kcu.column_name,', ') as key_columns
+from information_schema.table_constraints tco
+join information_schema.key_column_usage kcu 
+     on kcu.constraint_name = tco.constraint_name
+     and kcu.constraint_schema = tco.constraint_schema
+     and kcu.constraint_name = tco.constraint_name
+where tco.constraint_type = 'PRIMARY KEY'
+group by tco.constraint_name,
+       kcu.table_schema,
+       kcu.table_name
+order by kcu.table_schema,
+         kcu.table_name;
