--- server/test.postman_collection.json
@@ -1,37 +1,0 @@
-{
-	"info": {
-		"_postman_id": "73176cb3-05b0-49b2-aefa-4c8727f56b08",
-		"name": "test",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "http://localhost:3000/getAllRates",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"curr1\": \"USD\",\r\n    \"curr2\": \"EUR\",\r\n    \"basecurr\": \"USD\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/getAllRates",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"getAllRates"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
